Replace inheritance with a HOC
This commit will replace the Listenable superclass with a WithListeners
higher-order component.
