+++ packages/components/src/index.js
@@ -1,13 +1,16 @@
 'use strict';
 
-import Tracker from './viewport/tracker';
-import Aware from './viewport/aware';
+import ViewportTracker from './viewport/tracker';
+import ViewportAware from './viewport/aware';
 import WithPlaceholder from './viewport/withPlaceholder';
+import WithListeners from './shared/withListeners';
+
+const ViewportTrackerWithListeners = WithListeners(ViewportTracker);
 
 const Viewport = {
-  Tracker,
-  Aware,
-  WithPlaceholder,
+  Tracker: ViewportTrackerWithListeners,
+  Aware: ViewportAware,
+  WithPlaceholder: WithPlaceholder,
 };
 
 export {
